app-assankheti-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
motor
pydantic
pydantic-settings
pyjwt
passlib[bcrypt]
python-multipart
email-validator
python-jose[cryptography]
stytch
numpy
tensorflow==2.19.0
Pillow
beautifulsoup4
requests
